Start of the c++ implementation for coefficients
Includes a linked list for solutions, ways to modify solutions in the linked list, a class for rational numbers, a function to get the first digit of a number, a binomial coefficient number that returns its solution in rational form. The foundation for the rest of the c++ implementation for the Laurent Coefficients, still needs a lot more work on it.